requirements.txt
langchain==0.0.209
openai
streamlit
streamlit-chat
python-dotenv